Add sponsor to footer
Michigan Rec Sports requires their logo on our website, so it's added to
the bottom of the footer in the `sponsor.html` partial.

The logic for the partial is somewhat complex, since only one element is
needed to activate it and the image must be displayed on both root and
non-root pages.

Signed-off-by: Ethan Madison <ethan@ethanmad.com>