Replace $isHomePage with IsHome variable
A recent version of Hugo implemented the '.IsHome' page variable, which
replaces the now broken $isHomePage. (Hugo v1.8 broke title comparison
to determine if the current page is the home page since the home page's
title is "_index").

This commit replaces the broken $isHomePage with the functional .IsHome
and cleans up the logic used around using .IsHome.

Signed-off-by: Ethan Madison <ethan@ethanmad.com>