Add Welcome page
This page will replace the annual "Welcome to the Club" email that I've
sent out in early September the past two years. Much less typing!

Since there are now two pages about joining the club, there will now be
a "join" category. The Join and Welcome pages are part of this category.

Signed-off-by: Ethan Madison <ethan@ethanmad.com>